feat: create the crud service to reuse for model service to getList, create, update, getone and delete
