+++ src/controllers/property-type.controller.js
@@ -4,12 +4,15 @@ const PropertyTypeService = require("../services/property-type.service");
 const getPropertiesType = async (req, res, next) => {
   const { filters, search, sort, page, limit } = req.query;
 
-  const propertyTypes = await PropertyTypeService.getAllPropertyTypes({
+  const propertyTypes = await PropertyTypeService.getList({
     filters,
     search,
     sort,
     page,
     limit,
+    options: {
+      searchDefault: ["name"],
+    },
   });
 
   return new SuccessResponse({
